third_party/blink/web_tests/external/wpt/css/css-anchor-position/anchor-name-style-contained.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] .target 1
  assert_equals: \n<div class="container">\n  <div class="anchor a1"></div>\n  <div class="contain-style">\n    <div class="anchor a1"></div>\n  </div>\n  <div id="target-a1" class="target" data-offset-y="100"></div>\n</div>\noffsetTop expected 100 but got 200
[FAIL] .target 2
  assert_equals: \n<div id="target-a2" class="target" data-offset-y="50"></div>\noffsetTop expected 50 but got 200
Harness: the test ran to completion.
